+++ src/renderable/renderable.js
@@ -478,9 +478,9 @@ import Color from "./../math/color.js";
      * @returns {Renderable} Reference to this object for method chaining
      */
     rotate(angle, v) {
-        if (!isNaN(angle)) {
+        if (angle !== 0) {
             this.currentTransform.rotate(angle, v);
-            //this.updateBoundsPos(this.pos.x, this.pos.y);
+            this.updateBounds();
             this.isDirty = true;
         }
         return this;
@@ -528,9 +528,18 @@ import Color from "./../math/color.js";
      * @returns {Bounds} this shape bounding box Rectangle object
      */
     updateBounds() {
-        super.updateBounds();
-        this.updateBoundsPos(this.pos.x, this.pos.y);
-        return this.getBounds();
+        var bounds = this.getBounds();
+
+        bounds.clear();
+        bounds.addFrame(
+            0,
+            0,
+            this.width,
+            this.height,
+            this.currentTransform
+        );
+        this.updateBoundsPos(this.pos.x + bounds.x, this.pos.y + bounds.y);
+        return bounds;
     }
 
     /**
@@ -543,10 +552,9 @@ import Color from "./../math/color.js";
          bounds.shift(newX, newY);
 
          if (typeof this.anchorPoint !== "undefined" && bounds.isFinite()) {
-             bounds.translate(
-                 -(this.anchorPoint.x * bounds.width),
-                 -(this.anchorPoint.y * bounds.height)
-             );
+            var ax = bounds.width * this.anchorPoint.x,
+                ay = bounds.height * this.anchorPoint.y;
+            bounds.translate(-ax, -ay);
          }
 
          /*