+++ tests/browser/spec/entity-spec.js
@@ -43,16 +43,6 @@ describe("Entity", function () {
         })).toEqual(true);
     });
 
-    /*
-     * XXX: Disabled until #580 is fixed:
-     * https://github.com/melonjs/melonJS/issues/580
-     */
-    xit("has the correct entity geometry: A", async () => {
-        expect(await page.evaluate(() => {
-            return (entity.pos.x === 100 && entity.pos.y === 100 && entity.width === 42 && entity.height === 74);
-        })).toEqual(true);
-    });
-
     it("has a second shape", async () => {
         await page.evaluate(() => {
             entity.body.addShape(defaultRectShape.clone().setShape(-10, -10, 32, 64));
@@ -78,7 +68,7 @@ describe("Entity", function () {
     it("has the correct entity bounds: B", async () => {
         expect(await page.evaluate(() => {
             var bounds = entity.getBounds();
-            return (bounds.x === 0 && bounds.y === 0 && bounds.width === 42 && bounds.height === 74);
+            return (bounds.x === -10 && bounds.y === -10 && bounds.width === 42 && bounds.height === 74);
         })).toEqual(true);
     });